Move Notification Result and enum to notification_service.dart
